Display description, images and map on venue pages
- Use mapbox-gl for mapping
- Use internal SmugImage for single venue image, multiple not yet supported
- Lots minor styling changes
- Rename long -> lon

+++ _plugins/smugmug.rb
@@ -39,7 +39,8 @@ class Smug
       cache_file.close
 
       if (not cache_data.key? "FetchTime" or
-        cache_data["FetchTime"] < cache_invalid_time) and api_key
+        cache_data["FetchTime"] < cache_invalid_time) and api_key and 
+        not ENV['SMUGMUG_CACHE_MAINTAIN']
         # Delete and do over as cache invalid
         Jekyll.logger.warn("SM cache invalidated:", "Refreshing #{id}")
         File.delete(cache_filename(id))
